New styling and feedback for the user
Styling of the boekenlijst cards changed, added a new indication to let the user know that its added to the boekenplank (removed the alert function).

Desktop styling edited to scale good to the size.

The logic behind the modal of the changeColor, is now changed to closing the modal after the user presses "Verzenden" (also removed the alert function here).